requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
flask-jwt-extended==4.7.1
flask-cors==6.0.2
psycopg[binary]==3.2.12
python-slugify==8.0.4
python-dotenv==1.1.1
